+++ project/application/GameObject/Edamame/EdamameTrivia.h
@@ -7,7 +7,7 @@ class EdamameTrivia
 private:
     SoundData triviaVoice_;
     std::array<std::string, 8> strings_;
-    uint32_t triviaNum_ = 0;
+    size_t triviaNum_ = 0;
 public:
     EdamameTrivia();
     ~EdamameTrivia();
